- added parameters hook type: Vec2, Vec3, ChunkPos, BlockSource, CompoundTag

+++ src/native/main/core/mcpe/api/BlockPos.cpp
@@ -1,10 +1,9 @@
 #include <horizon/types.h>
 #include <java.h>
 #include <innercore_callbacks.h>
-#include <logger.h>
 
 export(jlong, mcpe_api_BlockPos_newClass, jint x, jint y, jint z){
-    return (jlong) new BlockPos(x, y, z);
+    return (jlong) new BlockPos((int) x, (int) y, (int) z);
 }
 export(void, mcpe_api_BlockPos_setX, jlong ptr, jint x){
     ((BlockPos*) ptr)->x = (int) x;
